camera: store depth as float32 (halves a depth-enabled episode)
get_depth() cast to float64, which was pure overhead: SAPIEN's "Position"
render target is float32 to begin with, so widening it added no precision
while doubling the bytes -- and depth is essentially the whole file on a
depth-enabled config (measured on demo_clean_multimodal: 1.17 GB of a 1.21 GB
episode, four cameras at 240x320 over 477 frames). float32 still resolves
~0.0002 mm at metre-scale ranges, far finer than the sensor model, and it is
what CLAUDE.md §7b already documented the column as.

Per episode ~1.21 GB -> ~0.63 GB, so a 50-task x 10-episode run drops from
~575 GB to ~312 GB, which is what makes it fit the remaining /project quota.

The in-place alpha mask (`depth *= rgba[...,3]/255`) keeps the narrower dtype
-- numpy in-place ops cast the float64 RHS down -- so the stored column is
float32 end to end.

CLAUDE.md §3.4 gains the depth-size note and, separately, records that every
entry in the embodiment's static_camera_list is captured unconditionally:
collect_head_camera gates only head_camera, so aloha-agilex's front_camera is
collected too, a quarter of the depth bytes above, with no data_type switch
for it.

+++ envs/camera/camera.py
@@ -444,15 +444,22 @@ class Camera:
     # Get Camera Depth
     def get_depth(self) -> dict:
 
+        # Depth is stored float32, in millimetres. float64 was pure overhead: SAPIEN's
+        # "Position" render target is itself float32, so widening the cast added no
+        # precision at all while doubling the bytes -- and depth dominates a
+        # depth-enabled config's on-disk size (measured: 1.17 GB of a 1.21 GB episode,
+        # four cameras at 240x320x477 frames). float32 still resolves ~0.001 mm at
+        # metre-scale ranges, far finer than the sensor model. This also matches what
+        # CLAUDE.md §7b already documents the column as.
         def _get_depth(camera):
             position = camera.get_picture("Position")
             depth = -position[..., 2]
-            depth_image = (depth * 1000.0).astype(np.float64)
+            depth_image = (depth * 1000.0).astype(np.float32)
             return depth_image
 
         def _get_sensor_depth(sensor):
             depth = sensor.get_depth()
-            depth = (depth * 1000.0).astype(np.float64)
+            depth = (depth * 1000.0).astype(np.float32)
             return depth
 
         res = {}
